+++ cpp/FMU-proxy/src/fmi/FmuHelper.hpp
